Move "content" to "destiny1", extract routes
